Update tips.txt
Merged the "Seemingly random debuffs can stem from a bad diet." with the "Beware the effects of poor nutrition." into "Beware the effects of poor nutrition. Seemingly random debuffs can stem from a bad diet." Neither hints are wrong, but they're practically the same hint. This change isn't needed, but I think it's a good merge.

Similarly, the two AE2 hints on line 59 and 60 could be merged, but merging the two as they are would make it far too long.